// apps/web/src/api/generated/model/apiKeyDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Open WebUI Klon API
 * OpenAPI spec version: 0.0.0
 */

export interface ApiKeyDto {
  /** @nullable */
  expiresAt: string | null;
  /** @nullable */
  lastUsedAt: string | null;
  id: string;
  name: string;
  prefix: string;
  createdAt: string;
}